+++ src/ru/topjava/webapp/storage/AbstractArrayStorage.java
@@ -1,6 +1,5 @@
 package ru.topjava.webapp.storage;
 
-import ru.topjava.webapp.exception.StorageException;
 import ru.topjava.webapp.model.Resume;
 
 import java.util.Arrays;
@@ -9,9 +8,8 @@ import java.util.Arrays;
  * Array based storage for Resumes
  */
 public abstract class AbstractArrayStorage extends AbstractStorage {
-    protected static final int STORAGE_LIMIT = 10_000;
+
     protected Resume[] storage = new Resume[STORAGE_LIMIT];
-    protected int size = 0;
 
     @Override
     public int size() {
@@ -19,7 +17,7 @@ public abstract class AbstractArrayStorage extends AbstractStorage {
     }
 
     @Override
-    protected Resume getResumeForIndex(int index) {
+    protected Resume getResume(int index) {
         return storage[index];
     }
 
@@ -30,19 +28,12 @@ public abstract class AbstractArrayStorage extends AbstractStorage {
     }
 
     @Override
-    protected void checkSizeToLimitArray(String uuidRes) {
-        if (size >= STORAGE_LIMIT) {
-            throw new StorageException("Хранилище уже заполнено - резюме невозможно сохранить!", uuidRes);
-        }
-    }
-
-    @Override
     public final void updateResumeToStorage(Resume resume, int index) {
         storage[index] = resume;
     }
 
     @Override
-    public final void deleteResumeAndTrimStorage(int index) {
+    public final void deleteResume(int index) {
         size--;
         deleteResumeFromStorage(index);
         storage[size] = null;
